Add SPHEREx logos to theme configuration
The logos are added to the new spherexsphinx/assets directory, and are
added to the Sphinx configuration in spherexsphinx.conf.base. The
get_asset_path helper function helps to provide the path to these
package assets for the Sphinx config. This method means that individual
projects don't need to include the logos.