Configure GitHub edit link
Add a SphinxConfig.set_edit_on_github method that configures the theme's
edit on GitHub functionality. The configuration uses the repository link
in the configuration TOML file in conjunction with tooling built on
GitPython to determine the path of the documentation project relative to
the root of the Git repository.